--- data/duke.txt
@@ -1,1 +1,0 @@
-E | Not Done | project meeting
